plugins/cni: add installation script for policy enforcement mode
When POLICY_ENFORCEMENT_INTERFACE is set, the cni configuration that
chains flannel with Cilium will be installed and uninstalled.

+++ plugins/cilium-cni/cni-install.sh
@@ -3,7 +3,15 @@
 set -eu
 
 HOST_PREFIX=${HOST_PREFIX:-/host}
-CNI_CONF_NAME=${CNI_CONF_NAME:-05-cilium.conf}
+if [ -z "${CILIUM_POLICY_ENFORCEMENT_INTERFACE}" ]; then
+    CNI_CONF_NAME=${CNI_CONF_NAME:-05-cilium.conf}
+else
+    until ip link show "${CILIUM_POLICY_ENFORCEMENT_INTERFACE}" &>/dev/null ; do
+        echo "Waiting for ${CILIUM_POLICY_ENFORCEMENT_INTERFACE} to be initialized"
+        sleep 1s
+    done
+    CNI_CONF_NAME=${CNI_CONF_NAME:-04-cilium.conflist}
+fi
 MTU=${MTU:-1500}
 
 BIN_NAME=cilium-cni
@@ -37,12 +45,40 @@ if [ -f "${CILIUM_CNI_CONF}" ]; then
 	echo "Using existing ${CILIUM_CNI_CONF}..."
 else
 	echo "Installing new ${CILIUM_CNI_CONF}..."
-	cat > ${CNI_CONF_NAME} <<EOF
+    if [ -z "${CILIUM_POLICY_ENFORCEMENT_INTERFACE}" ]; then
+        cat > ${CNI_CONF_NAME} <<EOF
 {
     "name": "cilium",
     "type": "cilium-cni"
 }
 EOF
+    else
+        cat > ${CNI_CONF_NAME} <<EOF
+{
+  "cniVersion": "0.3.1",
+  "name": "flannel-cilium",
+  "plugins": [
+    {
+      "type": "flannel",
+      "delegate": {
+         "hairpinMode": true,
+         "isDefaultGateway": true
+      }
+    },
+    {
+      "type": "portmap",
+      "capabilities": {
+        "portMappings": true
+      }
+    },
+    {
+       "name": "cilium",
+       "type": "cilium-cni"
+    }
+  ]
+}
+EOF
+    fi
 	if [ ! -d $(dirname $CILIUM_CNI_CONF) ]; then
 		mkdir -p $(dirname $CILIUM_CNI_CONF)
 	fi
